Extract utils to utilities/pinecone_utils.py
